+++ StudyManagement.Domain/OrderAgg/Order.cs
@@ -5,13 +5,13 @@ namespace StudyManagement.Domain.OrderAgg
     public class Order : EntityBase
     {
         public long AccountId { get; private set; }
-        public int TotalAmount { get; private set; }
+        public long TotalAmount { get; private set; }
         public bool IsPayed { get; private set; }
         public string IssueTrackingNo { get; private set; }
         public long RefId { get; private set; }
         public List<OrderItem> Items { get; private set; }
 
-        public Order(long accountId, int totalAmount)
+        public Order(long accountId, long totalAmount)
         {
             AccountId = accountId;
             TotalAmount = totalAmount;
