+++ StudyManagement.Domain/OrderAgg/OrderItem.cs
@@ -5,7 +5,7 @@ namespace StudyManagement.Domain.OrderAgg;
 public class OrderItem : EntityBase
 {
     public long SessionId { get; private set; }
-    public int SessionPrice { get; private set; }
+    public long SessionPrice { get; private set; }
     public int SessionNumber { get; private set; }
     public long ProfessorId { get; private set; }
     public string ProfessorFullName { get; private set; }
@@ -18,7 +18,7 @@ public class OrderItem : EntityBase
     public Order Order { get; private set; }
 
 
-    public OrderItem(long sessionId, int sessionPrice, int sessionNumber,long professorId, string professorFullName, string classStartTime, string classEndTime, int classDay, long classId, string courseName)
+    public OrderItem(long sessionId, long sessionPrice, int sessionNumber,long professorId, string professorFullName, string classStartTime, string classEndTime, int classDay, long classId, string courseName)
     {
         SessionId = sessionId;
         SessionPrice = sessionPrice;
